ConfigurationTest: Extract Schema verification functions and properties to new TestData class
